requirements.txt
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.14.0
greenlet==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
platformdirs==4.2.1
psycopg2-binary==2.9.9
pymssql==2.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==2.0.7
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
scramp==1.4.5
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
Werkzeug==3.0.2
zeep==4.2.1
google-cloud-storage==1.42.0